+++ errors/index.ts
@@ -3,3 +3,4 @@ export * from "./BloumeChatAPIError";
 export * from "./BloumeChatAuthError";
 export * from "./BloumeChatTimeoutError";
 export * from "./RateLimitError";
+export * from "./BloumeChatVoiceError";